+++ oauth2/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Made code async using `tokio`.
+
 ## [0.0.3] - 2023-06-06
 
 ### Added